Moved stop place queries to separate folder. Added util for journey planner queries
