+++ .flowconfig
@@ -8,6 +8,6 @@ flow-typed
 [lints]
 
 [options]
-
+esproposal.optional_chaining=enable
 
 [strict]
