+++ src/bikeRental/index.js
@@ -1,20 +1,16 @@
 // @flow
-import { post } from '../api'
+import { journeyPlannerQuery } from '../api'
+
 import { getBikeRentalStationProp, getBikeRentalStationByBoxProps } from './query'
-import { getJourneyPlannerHost } from '../config'
 import { convertPositionToBbox } from '../utils'
 import type { BikeRentalStation, Coordinates } from '../../flow-types'
 
 export function getBikeRentalStation(stationId: string): Promise<BikeRentalStation> {
-    const { host, headers } = getJourneyPlannerHost(this.config)
-
-    const url = `${host}/graphql`
     const variables = {
         id: stationId,
     }
-    const params = { query: getBikeRentalStationProp, variables }
 
-    return post(url, params, headers)
+    return journeyPlannerQuery(getBikeRentalStationProp, variables, this.config)
         .then(response => response.data.bikeRentalStation)
 }
 
@@ -22,12 +18,8 @@ export function getBikeRentalStations(
     coordinates: Coordinates,
     distance: number = 500,
 ): Promise<Array<BikeRentalStation>> {
-    const { host, headers } = getJourneyPlannerHost(this.config)
-    const url = `${host}/graphql`
-
     const variables = convertPositionToBbox(coordinates, distance)
-    const params = { query: getBikeRentalStationByBoxProps, variables }
 
-    return post(url, params, headers)
-        .then(response => (response.data || {}).bikeRentalStationsByBbox || [])
+    return journeyPlannerQuery(getBikeRentalStationByBoxProps, variables, this.config)
+        .then((response: Object = {}) => response?.data?.bikeRentalStationsByBbox || [])
 }
