+++ src/trip/index.js
@@ -1,24 +1,19 @@
 // @flow
-import { post } from '../api'
+import { journeyPlannerQuery } from '../api'
 import {
     FOOT, BUS, TRAM, RAIL, METRO, WATER, AIR,
 } from '../constants/travelModes'
 import {
     getItinerariesProps,
     getStopPlaceDeparturesProps,
-    getStopPlaceProps,
-    getStopPlacesByBboxProps,
 } from './query'
-import { getJourneyPlannerHost } from '../config'
 
 import type {
-    Coordinates,
     TripPattern,
     Location,
-    StopPlace,
     LegMode,
 } from '../../flow-types'
-import { convertPositionToBbox, convertFeatureToLocation, isValidDate } from '../utils'
+import { convertFeatureToLocation, isValidDate } from '../utils'
 
 type StopPlaceParams = {
     onForBoarding?: boolean, // deprecated
@@ -50,9 +45,6 @@ export type GetTripPatternsParams = {
     wheelchairAccessible?: boolean,
 }
 export function getTripPatterns(searchParams: GetTripPatternsParams): Promise<Array<TripPattern>> {
-    const { host, headers } = getJourneyPlannerHost(this.config)
-    const url = `${host}/graphql`
-
     const {
         searchDate, limit, wheelchairAccessible, ...rest
     } = { ...DEFAULT_SEARCH_PARAMS, ...searchParams }
@@ -64,14 +56,8 @@ export function getTripPatterns(searchParams: GetTripPatternsParams): Promise<Ar
         wheelchair: wheelchairAccessible,
     }
 
-    return post(url, { query: getItinerariesProps, variables }, headers)
-        .then((response: Object) => {
-            try {
-                return response.data.trip.tripPatterns
-            } catch (e) {
-                return []
-            }
-        })
+    return journeyPlannerQuery(getItinerariesProps, variables, this.config)
+        .then((response: Object = {}) => response?.data?.trip?.tripPatterns || [])
 }
 
 export async function findTrips(
@@ -109,7 +95,6 @@ export function getStopPlaceDepartures(
     stopPlaceIds: string | Array<string>,
     stopPlaceParams?: StopPlaceParams,
 ): Object {
-    const { host, headers } = getJourneyPlannerHost(this.config)
     const {
         timeRange, departures, onForBoarding, includeNonBoarding,
     } = { ...DEFAULT_STOP_PLACE_PARAMS, ...stopPlaceParams }
@@ -122,8 +107,6 @@ export function getStopPlaceDepartures(
         omitNonBoarding = !onForBoarding
     }
 
-    const url = `${host}/graphql`
-
     const askingForSingleStopPlace = typeof stopPlaceIds === 'string'
 
     const variables = {
@@ -134,10 +117,8 @@ export function getStopPlaceDepartures(
         omitNonBoarding,
     }
 
-    const params = { query: getStopPlaceDeparturesProps, variables }
-
-    return post(url, params, headers)
-        .then((response: Object) => {
+    return journeyPlannerQuery(getStopPlaceDeparturesProps, variables, this.config)
+        .then((response: Object = {}) => {
             if (!response || !response.data) {
                 throw new Error(`Entur SDK: Could not fetch departures for ids: ${JSON.stringify(stopPlaceIds)}`)
             }
@@ -149,42 +130,5 @@ export function getStopPlaceDepartures(
                 id,
                 departures: estimatedCalls,
             }))
-        })
-}
-
-export function getStopPlace(id: string): Promise<StopPlace> {
-    const { host, headers } = getJourneyPlannerHost(this.config)
-    const url = `${host}/graphql`
-
-    const variables = { id }
-    const params = { query: getStopPlaceProps, variables }
-
-    return post(url, params, headers)
-        .then((response) => {
-            try {
-                return response.data.stopPlace
-            } catch (e) {
-                throw new Error(`Could not find stop place with ID "${id}"`)
-            }
-        })
-}
-
-export function getStopPlacesByPosition(
-    coordinates: Coordinates,
-    distance: number = 500,
-): Promise<Array<StopPlace>> {
-    const { host, headers } = getJourneyPlannerHost(this.config)
-    const url = `${host}/graphql`
-
-    const variables = convertPositionToBbox(coordinates, distance)
-    const params = { query: getStopPlacesByBboxProps, variables }
-
-    return post(url, params, headers)
-        .then((response) => {
-            try {
-                return response.data.stopPlacesByBbox
-            } catch (e) {
-                return []
-            }
         })
 }
